Rename components for better understanding
