+++ app/containers/TicketTab/TicketTab.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
-import TicketPage from 'components/TicketPage';
+import TicketTab from 'components/TicketTab';
 import { actions, selectors } from 'reducers/ticket';
 import { compose } from 'redux';
 
@@ -12,4 +12,4 @@ const mapDispatchToProps = {
   getAllAction: actions.getAllAction,
 };
 
-export default compose(withRouter, connect(mapStateToProps, mapDispatchToProps))(TicketPage);
+export default compose(withRouter, connect(mapStateToProps, mapDispatchToProps))(TicketTab);
