--- app/pages/TicketPage/index.js
@@ -1,1 +1,0 @@
-export { default } from '../../containers/TicketPage';
